+++ packages/components-js/tests/e2e/specs/accordion.e2e.ts
@@ -1,5 +1,5 @@
-import type { Page } from 'playwright';
 import { expect, test } from '@playwright/test';
+import type { Page } from 'playwright';
 import {
   addEventListener,
   getAttribute,
@@ -243,6 +243,19 @@ test.describe('focus', () => {
     await waitForStencilLifecycle(page);
 
     expect(await hasFocus(body)).toBe(true);
+  });
+
+  test('should set programmatic focus', async ({ page }) => {
+    await initAccordion(page);
+    const host = getHost(page);
+
+    await expect(getButton(page)).not.toBeFocused();
+
+    await host.focus();
+    await expect(getButton(page)).toBeFocused();
+
+    await getButton(page).blur();
+    await expect(getButton(page)).not.toBeFocused();
   });
 });