+++ packages/components/CHANGELOG.md
@@ -14,12 +14,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+## [3.34.0] - 2026-04-09
+
+## [3.34.0-rc.0] - 2026-04-01
+
+### Added
+
+- `Segmented Control`: `noWrap` property to render items inline with optional scroller
+  ([#4244](https://github.com/porsche-design-system/porsche-design-system/pull/4244))
+
+### Fixed
+
+- `Accordion`: missing `delegatesFocus`
+  ([#4238](https://github.com/porsche-design-system/porsche-design-system/pull/4238))
+
+## [3.33.0] - 2026-03-09
+
 ## [3.33.0-rc.0] - 2026-03-04
 
 ### Added
 
 - `Tag`: new `variant` property to define background colors which complies now with PDS `v4`
   ([#4227](https://github.com/porsche-design-system/porsche-design-system/pull/4227))
+- `Checkbox`: add warning for Firefox form restore bug
 
 ### Changed